SmartWaste/requirements.txt
flask
joblib
numpy
Pillow
scikit-learn
